fix: set fallback texture name after the struct assignment
If the `kstrncp` function is called before the `texture` struct
assignment to `state->fallback_texture`, when the assignment is done,
the `name` field gets discarted by a `NULL` (falsy) value. Thus, the
engine was stating that the fallback texture had no name when wanting
to destroy it.

When calling the `kstrncp` function to set the `name` field is done
after the `texture` struct assignment, this problem gets avoided.