fix: `kstrsub` function array wrong memory access (stack smashing)
To solve an error in the `kstrsub` function of the `kstring` module,
this patch is being submitted.
The memory access done in `dest[start + len] = 0;` is not correct, as
the wanted substring is `len` long; thus, it is only needed to
terminate the selected substring in `dest` after the amount of
characters specified through the `len` argument.